Reworked Cell class, added text fields, added images to Cells depending on owner.
Added a few text fields to the board, so the player can see how
many wins for players x and o, and how many ties have been had.
Made Cells extend JLabels, so that rather than having meaningless
JLabels on the board, I now have Cells with associated graphics.
And now, when the human player clicks a Cell, ownership of that
Cell is passed to him/her (player X).