Implemented end game dialog, AI rolling, and winner checking
Whenever a player wins or there is a tie game, now a dialog pops
up telling the player who won.
The computer player now will automatically roll and select a
random cell after the player moves.
Also, after each game, the total number of scores is updated
on the right side of the screen.