More code cleanup - see extended description.
Replaced the createControl() method with a new method
addComponent(), that takes a more generalized JComponent (rather
than just a JButton) as an argument. Was able to apply this
method to several more GUI objects.
Simplified the constructor for Cells a bit.
Changed the variables  "roll" to "randomRoller" and "num"
to "rollValue" for clarity.